requirements.txt
appdirs==1.4.3
asn1crypto==0.22.0
attrs==16.3.0
Automat==0.5.0
certifi==2017.1.23
cffi==1.10.0
constantly==15.1.0
cryptography==1.8.1
cssselect==1.0.1
enum34==1.1.6
future==0.16.0
idna==2.5
incremental==16.10.1
ipaddress==1.0.18
lxml==3.7.3
packaging==16.8
parsel==1.1.0
pbr==1.8.1
psycopg2==2.7.1
pyasn1==0.2.3
pyasn1-modules==0.0.8
pycparser==2.17
PyDispatcher==2.0.5
pyOpenSSL==17.0.0
pyparsing==2.2.0
pytidylib==0.3.2
queuelib==1.4.2
Scrapy==1.3.3
service-identity==16.0.0
six==1.10.0
stevedore==1.12.0
Twisted==17.1.0
virtualenv==15.1.0
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.1
w3lib==1.17.0
zope.interface==4.4.0
